omg-cli: Implement google project, consume it in prepare-project

+++ src/omg-cli/omg/setup/gcp_project.go
@@ -30,7 +30,7 @@ type ProjectValidator struct {
 
 type QuotaError struct {
 	google.Quota
-	Actual float32
+	Actual float64
 }
 
 var UnsatisfiedQuotaErr = errors.New("quota unsatisfied")
